requirements.txt
Flask==2.1.2
flask-restx==0.5.1
flask-cors==3.0.10
elasticsearch==8.2.0
werkzeug==2.1.2
python-dateutil==2.8.2
pandas==1.4.2
kafka-python==2.0.2
gunicorn==20.1.0
gevent==21.12.0
requests ==2.26.0